+++ examples/PyMPDATA_examples/comparison_against_pypde_2025/diffusion_2d.py
